Move version-oriented functions to their own file
